Moved some tests to test_config to check for correct exception
